# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Library API",
    "description": "CRUD API for managing Books and Authors collections in MongoDB. Write operations (POST, PUT, DELETE) are protected and require the user to be logged in via GitHub OAuth (use /login in a browser tab first, since Swagger's 'Execute' shares your browser's session cookie).",
    "version": "1.1.0"
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Local server" },
    { "url": "https://library-api-1wil.onrender.com", "description": "Render server" }
  ],
...
  "paths": {
...
    "/auth/github/callback": {
      "get": {
        "summary": "GitHub OAuth callback",
        "tags": ["Auth"],
        "description": "GitHub redirects here automatically after the user authorizes the app. Not meant to be called directly.",
        "responses": {
          "200": { "description": "Logged in successfully, session created" },
          "302": { "description": "Redirected back on failure" }
        }
      }
    },
...
  },
...
}